+++ resdk/__init__.py
@@ -1,5 +1,6 @@
 """Resolwe SDK for Python."""
 from __future__ import absolute_import, division, print_function, unicode_literals
 
+from .analysis import _register  # Patch ReSDK resources with analysis functions
 from .resolwe import Resolwe, ResolweQuery
 from .resdk_logger import log_to_stdout, start_logging