+++ resdk/resolwe.py
@@ -30,7 +30,6 @@ from .resources import Data, Collection, Sample, Process, Relation
 from .resources.kb import Feature, Mapping
 from .resources.utils import (
     iterate_fields, iterate_schema, endswith_colon, get_collection_id, get_data_id)
-from .utils.resolwe import ResolweUtilsMixin
 from .query import ResolweQuery
 
 
@@ -68,7 +67,7 @@ class ResolweAPI(slumber.API):
     resource_class = ResolweResource
 
 
-class Resolwe(ResolweUtilsMixin):
+class Resolwe(object):
     """Connect to a Resolwe server.
 
     :param email: user's email